api/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-bigquery
google-cloud-storage
pydantic
db-dtypes
pandas
numpy
gunicorn
